Add LiveFunctionsInterface
With this change, we add `LiveFunctionsInterface`. We made
`LiveFunctionsController` to inherit from `LiveFunctionsInterface` and
change `LiveFunctionsDataView` to use the interface class.

The purpose is to make `LiveFunctionsDataView` independent with the
OrbitGl module such that it can be moved to the DataViews module.
`LiveFunctionsDataView` only need the `AddIterator` method from
`LiveFunctionsController`, hence we can make a interface to provide that
functionality.