Implement Introspection using Manual Instrumentation Events.
This removes introspection from all events and all "client" specific
code for it, and instead use the new manual instrumentation events.

Collecting the events is still handled differently
(using TracingListeners instead of the injection of liborbit.so),
but all protos used, as well as all code that handles the events
is the same as for manual instrumentation.

This allows us to simplify a lot of code and remove all the special
handling.

Bug: http://b/196032048 http://b/182848566
Test: Take a capture on OrbitTest and --devmode, start introspection