Add tests for LiveFunctionsDataView
With this change, we add tests for `LiveFunctionsDataView`.

As the view needs to get function statistic values from `FunctionStats`
by calling `app_->GetCaptureData().GetOrCreateFunctionStats()` and there
is no easy way to add mock function stats to `CaptureData`, we created
MockCaptureData, which is a mock of CaptureData providing mock methods
to get FunctionStats and determine frame track is enabled or not.