Unify Track::GetHeight and GetSize
Before this commit, we used to have both methods and we needed to sync
them before every draw call. Now, GetSize returns a pair of width and
height using GetHeight().

In adition, SetWidth now also set width of each children in Gpu and
ThreadTrack. In a future PR, we will want to make this generic.
Now we are also asking for an update when modifying this method.

There were a few bug related with the syncing, a current one is
(http://b/196198708) about VerticalZoom and Collapsing/Expanding Tracks.

Bug related: http://b/184245377.