Split UpdatePrimitives with size calculation
This commit is a bunch of small commits which are all together. Sorry
for that, I'll try to list everything here.

- Revert the hack for drawing twice when needed (viewport_was_dirty -
https://github.com/google/orbit/pull/2209). Not needed anymore.
- Split GetSize/GetPos with drawing methods. First update, after draw.
(http://b/187502610).
- Move right_margin from TimeGraph to CaptureWindows.
(http://b/192070555). Also simplify how we update it.
- Make a method IsVisible for Vertical Slider.
- Make a method to calculate the visible tracks total height. (Fixed
issues with outdated size of slider).
- Refactor in TrackManager/CaptureWindows (we had several issues here with
visible_tracks). This is motivated with the long list of problems here.
Now we have a clear order with sorting the tracks and update the visible
track list. Every function ask and it is updated in the first step
before starting to draw neither calculating positions.
- SetPos in tracks is the only who request an update of subtracks position.
- We fix a test in TrackManager

As a result, http://b/196198708 is also fixed.

Test: running capture, loaded capture.

As a future plan we will want:
- Generalize SetWidth/SetPos to include childrens (with the future interface of
tracks and panes).

Revert invariant of visible track list in Track Manager

We are having issues with Unit tests. This invariant force to have a
visible track lists always updated but when we create the tracks, they
are empty and therefore they won't be visible. Later a timer appears but
TrackManager doesn't know that the lists of tracks needs an update.

This will be solve in the next SortTracks, but it's too late for Unit
Tests.

So, we are reverting this invariant and instead forcing to update Tracks
before updating sliders.