Don't build missing packages on the CI
The CI used to build missing dependency packages automatically. This is
problematic because the user won't notice that they missed uploading a
prebuilt package to our artifactory servers (both internal and
external).

But sometimes you want to have this automatic behaviour. In this case it
will be possible to manually trigger a CI job an set the environment
variable `ORBIT_BUILD_MISSING_PACKAGES` there.