Implement build id for PE/COFF
PE/COFF object files and corresponding debug info in PDB files are
uniquely identifid by GUID and age in the PDB info found in object
and PDB files. We form a build id that we use to identify modules by
concatenating the GUID and age in a human readable string.

Tested: Unit test.
Bug: http://b/194497774