Move GetPos/GetSize to derived classes
To avoid having the if inside the methods GetPos and GetSize and having
mock methods for get them or GetBarPixelLength.

Also, seems better to have a better name inside the unit tests, so I
changed it there to have UnitTestHorizontalSlider and
UnitTestVerticalSlider instead of having them hiding in a parameter.

I'm open to suggestions here.