Consistently use GetWidth() to fix disappeared timers
In
https://github.com/google/orbit/commit/44ec55c7942ef1dae7a6f6363451c40e0ce3e85e,
we started to use GetWidth() instead of viewport->ScreenWidth() for the
width of the tracks, since we were no longer considering the vertical
slider.

Also, since we are now using GetWidth(), we are previously using
viewport method's WorldToScreenWidth to convert to screen space.

However there were still a few missing references we didn't change, which
caused an strange behavior in ThreadTrack. So, in this PR we erase them.

Test: Load a capture.

Pd: There are still a few references in AccessibleCaptureViewElement
which I'm not sure how to solve it.

fake commit