Use WorldToScreen conversion in the right line
Caused by lack of attention. Thanks Florian for pointing
it. Explained better in  https://github.com/google/orbit/pull/2695. We
were using GetScreenWidth from viewport, we wanted to use instead
timegraph->GetWidth() and for that we had to convert to Screen.

Now we are doing that in the right line. By the way, still not sure
about if we have to add in ToWorldX(), but before it wasn't and I will
leave as it was before. There are 4 occurences of the function, the
other 3 are not using it.