Fix wrongly starting position of slider
In this PR fix an issue related with wrongly starting position of a
slider.

The issue is caused by 3 things:
1- The slider initial size is wrong (20, instead of 15).
2- We have to update first one slider and after the other one.
3- The slider sets their size based on the orthogonal one.

The best solution in my opinion is addressing point 3 by using the
method GetSliderWidth() from TimeGraphLayout directly. As we don't have
this TimeGraph already created there, it is needed a refactor I won't
make here (and also probably renaming to CaptureWindowsLayout).

So, I only solved the issue by using directly this method in
CaptureWindows given the fact that both Sliders will have the same size.
Aditionally I'm modifying the starting size (point 2) but it's not
really needed.

http://b/197211318.