Add FuzzingUtils
The new implementation of `PLATFORM_ABORT` specific for fuzzing throws
instead of aborting. This allows a fuzzer to recover from a failed
CHECK. But it also means we need to catch the exception somewhere.

We can't just put a try-catch-block into the fuzzer's main function
because the fuzzer should still compile with `-fno-excepetions` enabled.

That's why I'm introducing two macros that wrap everything in a
try-catch-blocks if the fuzzing build mode is enabled.