Fix position in legend's tooltip
Regression from https://github.com/google/orbit/pull/2670.

We changed the position of the transparent box where we were drawing
the legend, and as it wasn't aligned with the text drawn, we were
showing an incorrect tooltip there.

Fixed it by simply moving together the text and box before modifying the
position.

https://b/197844515