Avoid ABSL_DECLARE_FLAG in Service
This removes ABSL_DECLARE_FLAG from OrbitGrpcServer.cpp and instead
passes in the value of the `devmode` flag via function and constructor
arguments similar to how it is already done with `grpc_port`.

This allows the `Service` target to be compiled as a shared library
without any missing symbols from Orbit.