Add "Enable callstack sampling" and "Callstack sampling period" options
They are added to the "Capture Options" dialog.
At least for now, they are only visible and only have effect in `--devmode`.

I chose to go for sampling period instead of sampling rate for consistency with
the already existing option to specify the memory sampling period.

Maximum period that can be specified is 1000 ms and minimum is 0.1 ms (it seems
that we are able to handle it).

Screenshot: http://screen/8kQich8AYGaFow3

Bug: http://b/160549506

Test: Capture Honolulu with different sampling periods in `--devmode`, and
verify behavior when not in `--devmode`.