Add GetTestLibLibraryPath function
This adds a common function that can be used in all
UserSpaceInstrumentation tests to obtain the location of the TestLib.

That's needed because the location and name is different in internal
builds. So this also introduces a copybara variable that will be
replaced for internal builds and allows to adjust the location.