Fail if test executable is not found in the list of modules
The test is looping through the list of modules and implicitly expects
to find a module with a hard-coded name.

If the name is not found the test fails in a weird way and it's hard for
the user to debug what's actually happening.

This change ensure the test immediately fails with a proper error
message when the module is not found.