Add interal path for libOrbitUserSpaceInstrumentation
The InstrumentProcess class looks into two locations
for the inject library. That's where the library can be found
during production and where it can be found when started from the build
directory (during development).

In internal builds the library is in a different place when tests run on
the CI so this commit is adding a third location when build internally.