Add new HasValueThat matcher and tests for matchers
This adds a matcher `HasValueThat` that takes a matcher for the stored
value as an argument.

This also adds an overload to the existing `HasValue` matcher which
behaves the same as the new `HasValueThat` matcher.

And since this is all getting more and more complex I added some tests
for the matchers.