Move CaptureServiceImpl to its own module
Part of a series of changes that aim to isolate each gRPC service into its own
module for better readability and maintenance. Currently, the Service module
contains multiple independent subsets of classes contributing to the different
services, so we extract those.

This makes it clearer what classes previously in the `Service` module play a
role in `CaptureServiceImpl` and which don't.
We probably need to considerably extend some of these classes soon (e.g.,
`ProducerCaptureEvent`) so this paves the way.

Test: Take a capture with various options enabled.