Remove --devmode flag for OrbitService
This was basically a leftover, so let's remove it for simplicity reasons.
Its only remaining usage was to register `CrashService`. But we can just always
register it, similarly to how we always register `TracepointService` and
`FramePointerValidatorService`. It will still only be possible to interact with
`CrashService` if the client was started in `--devmode`.

Test: Deploy `OrbitService` with `BareExecutableAndRootPasswordDeployment`,
start client in `--devmode`, verify it's still possible to crash `OrbitService`.