Delay searching CallTreeWidget to prevent searching on every character
Only search top-down and bottom-up views when we detect that the user has
finished typing, based on a timeout. As the trees can be quite big and
displaying the result of the search involves expanding and collapsing nodes,
searching on every new character typed can be laggy and interfere with writing
the full search term.

Bug: http://b/197499003

Test: Play around with searching "Top-Down" and "Bottom-Up", also across
multiple captures and multiple selections.