Move ScopeTree and BlockChain to Containers
We are moving all timer-related information to ClientData, to not have
any dependence between data and UI. We will avoid any include from data
to OrbitGl. As Blockchain is used for ScopeTree and has to be used in
ClientData but also is still being used in OrbitGl, we are moving both
classes to a new project named containers.

Bug we are focusing in: http://b/198135618