Use long for running deviation squared sum
There was an instance of an overflow when storing the running deviation
squared sum as a double in b/198279407. Switching it to an unsigned long
while doing the running calculation allows us to accommodate for larger
numbers and doesn't sacrifice precision as it can't be a decimal until
it gets cast.

Bug: http://b/198279407