Create ScopeTreeTimerData
We are creating ScopeTreeTimerData which will represent the timers inside
a ThreadTrack based on ScopeTree data structure. For now, it's a temporal
class because we are trying to query the timers for ThreadTrack. In the future
we will do the same for other TimerTracks and we might eventually merge it with
TrackPaneData.

As currently, we are having two containers, TrackPaneData and ScopeTree.
In the future we might use ScopeTree while capturing (with some
modifications to erase bunch of timers), and the other container will be
used when stopping a capture.

The current functions are AddTimer(), which will replace OnTimer and
ProcessTimer for ThreadTrack, BuildScopeTreeFromTrackData(), which will
be used to construct the ScopeTree when loading a capture (legacy), and
GetAllTimers(min_tick, max_tick) which will return all timers in that
range. Aditionally we have the methods used in TrackPaneData and
GetLeft/Right/Top/Down.