Fix address computation for symbols loaded from PDB files
Symbol addresses in PDB files are stored as so called "relative virtual
addresses", which is the offset from the start of the corresponding
.text section. Symbol addresses in ELF files however are stored as
"virtual addresses" which include the load bias and the executable
segment offset. To make all our address computations work consistently,
we pass in the load bias (ImageBase for COFF) and the executable
segment offset when loading the PDB file to adjust symbol address.

If the relative virtual addresses are desired when opening a PDB file,
one can simply pass 0, 0 as load bias and executable segment offset,
respectively.

Bug: http://b/194497774
Tested: Unit tests.