Fix number of visible primitives
We used to count the number of visible text boxes, but it currently says
0, so in this PR we renamed it (as TextBox don't exist anymore) and
counted using the number of visible primitives in every track.

Clarification: We are counting primitives of all the visible tracks
(including the ones which are not visible because are being drawn
outside the visible area), but this is a bug that should be fixed in
another PR).