Cache CallTreeView::children (#2784)
This is a huge speedup in both building and searching top-down and bottom-up
views, in particular the bottom-up view, and in particular when symbols for a
lot of functions are missing, which prevents grouping addresses belonging to the
same function together (which is a completely separate problem).

All credits to Nikita!

Also add a couple of `ORBIT_SCOPE_FUNCTION`+`SCOPED_TIMED_LOG`s, and fix a minor
clang-tidy warning in `SamplingReport`.

Bug: http://b/199712443

Test: Load the capture linked in the bug before and after and verify the huge
speedup. Take some more captures and verify the speedup and that top-down and
bottom-up views look good.