Only UpdatePrimitives and Draw tracks on screen
Optimization easy to make now that we are not setting
positions neither sizes while drawing anymore.

After calculation all the positions in WorldSpace we only
UpdatePrimitives and Draw to the tracks which have some part visible in
screen.

As a result, in captures where we have a lot of Tracks with timers
in several depths, we are much more faster at Updating Primitives. For
example, in the capture generated by Pierric using OrbitTest, with the
initial zoom I'm seeing UpdatePrimitives being 5/6 times faster, but
maybe could vary based on the resolution.

Anyway, in addition with https://github.com/google/orbit/pull/2780,
which is an optimization only for thread tracks, we are aiming to be
only bounded by the number of pixels where we draw timers multiplied by
a log of the number of elements.

Test: Loaded a capture, start a new one.