Make NsPerPixel a double instead of truncating
As mentioned in https://github.com/google/orbit/pull/2795, we are using
an integer value for ns_per_pixel, which leads us to a problem because
if we multiply the number of pixels with ns_per_pixel we have a
different value than the total number of pixels.

In this PR, we are switching it to double and we cast to an int for
getting the correct timestamp.

Test: Load a capture, see than the image is the same than the current
main.