Use Integers in GetNextPixelBoundaryTimeNs (#2795)
* Use Integers in GetNextPixelBoundaryTimeNs

We use the function GetNextPixelBoundaryTimeNs to get the next time to
draw, in order to not draw a timer which all its pixels have already
another element drawn.

This function have 2 issues which leads to return sometimes a previous
time (and therefore incorrect one), which is currently making us drawing
timers than otherwise we will skip.

The first issue is that we have an integer value for ns_per_pixel, which
leads us to a problem because if we multiply the number of pixel with
ns_per_pixel we don't have the total number of pixels. We will fix this
in a different PR. The second issue is that although we use a double
there, we will have precision issues. So, for that we will make all our
calculations using integers, to avoid that.

Addressing: http://b/199736246

Test: Navigate in a capture, confirm that always the function answer a
reasonable time. Load a capture, there were less elements being drawn (0.1%
less), but pixelwise the image is almost the same (I only saw an only
different line that it's more precise in the after-screenshot).

Before: https://screenshot.googleplex.com/76KYwkkjk7c3eup
After: https://screenshot.googleplex.com/4WEjtteXDethqMz

* Fixup: Assure that the new timestamp of a pixel is after the current one

* Fixup: Rename variables, comment and address a border case