Make vertical zoom point to the same element again (#2805)
* Make vertical zoom point to the same element again

Regression from 1.67 (https://github.com/google/orbit/pull/2670). The
function is not commented properly, name of the variables is not clear
enough, so wasn't updated during that refactoring.

The element being pointed with the mouse shouldn't change during
vertical-zooming. So, in this PR we are adapting the vertical zoom to
the new coordinate system and renaming some variables + adding a comment
to make that function easier to understand.

Fixing http://b/200505756.

Test: Load a capture and making a vertical zoom there.

* Fixup: Suggested renaming in comments