Optimize going through timers in ThreadTrack (#2780)
* Optimize going through timers in ThreadTrack

Here we are optimizing UpdatePrimitives in ThreadTracks by not going
through timers that will be drawn in the same pixel of another
previously drawn. We weren't drawing them anyways, but we were going
through all of them, making orbit very slow when there were a lot
of timers to draw.

My first experiment was with a capture of 38millions timers (hooking Dot function for 16.9 seconds), and
only drawing ThreadTracks shows that in median time TimeGraph::UpdatePrimitives was 39.4 times faster, making Orbit usable
in a capture with a lot of timers while zooming-out.
https://docs.google.com/spreadsheets/d/1Dm8HeCvVH1pMM7gk5oPJOLKyh3wNFj1T_tTSC-PjLf0/edit?resourcekey=0-4gLk_6bz1SXnBnKP4l-NTA#gid=0

Sadly, this is only the ideal case (a lot of timers in a few threads) and Pierric provided me a capture where the optimization seems to be closer to 10/20% (to be measured).

Speaking of complexity, where N is the number of timers in a certain depth level, D the number of drawn timers and T the extra time to draw an additional timer, we have:

Before: O(N) + O(D) * O(T)
Now: O(lg(N)) * O(D) + O(D) * O(T)

So, having a screen of 2000 pixels width, the optimization seems to be an improvement theoretically since having 30k visible timers in certain depth.

Orbit is being too slow anyways (far away from 30fps) and we are only drawing
ThreadTracks, so a future analysis should be made.

Test: Load a capture. Counted the number of timers drawn, the number doesn't
decrease with current state. Start a new capture.

* Fixup: Create FindFirstVisibleScopeAfterTime in ScopeTree

* Fixup: Simplify while loop

* Fixup: FindFirstScopeAtOrAfterTime that coming back to an if

* Fixup: Add a TODO for GpuTracks