Parse "Group Id" from encoded Vulkan debug labels.
In DXVK, we insert Vulkan debug labels with the following encoding:
'DXVK__vkFunctionName#GROUP_ID', where 'GROUP_ID' is the group
id to be used.

This parses the labels and sets the group id field if possible.

There are already tests for GpuQueueSubmissionProcessor in
the CaptureEventProcessorTest, but I believe it is better
to have those special tests closer to the actual class.

We should in the future, add some more tests for the corner
cases in the new test file.

Test: Take a capture on a DXVK sample, run tests.