Read on a protobuf Arena in ProducerSideServiceImpl, reading side of IPC (#2836)
We defer the memory allocations to when `ProducerEventProcessor::ProcessEvent`
is called. But the speedup of `ServerReaderWriter::Read`, otherwise particularly
inefficient without an Arena, greatly overcompensates and results in a big net
performance improvement.

I tried this with user space dynamic instrumentation and manual instrumentation
on a benchmark causing ~800k `FunctionCall`s per second, and ~800k pairs of
`ApiScopeStart`/`ApiScopeStop` per second, respectively.
In the former case the CPU utilization of the thread `PSSI::RcvEvents` (the one
that executes the aforementioned reads and copies) goes down by ~40% (relative)
in the former case (~38% to ~23%), and by ~33% (~75% to 50%) in the latter case.
The CPU utilization of the rest of `OrbitService` remains unchanged, so the
total CPU utilization of `OrbitService` adjusts accordingly to these numbers.
I also tried it on Trata's `dm_gfx::vulkan::vulkan_command_buffer::bind_set`
(~240kHz) with user space instrumentation. The improvements looked similar, but
the numbers fluctuated a lot more.

Bug: http://b/179969217

Test: See above.