Pass by rvalue-ref to ProducerEventProcessor::ProcessEvent
This normally doesn't make a difference, because internally a
`ProducerCaptureEvent` is (almost) only a pointer to the actual event:
- passing by value from the heap with a move is only a copy of the pointer;
- passing by value from a `protobuf::Arena` (in `ProducerSideServiceImpl`)
  causes a deep copy, but when passing by (rvalue) reference from a
  `protobuf::Arena` the deep copy will happen anyway when moving the actual
  event into the `ClientCaptureEvent`.

Where it does make a (big) difference is when passing from a `protobuf::Arena`
without simply moving into a `ClientCaptureEvent`, but for example storying some
fields in a different object. In this case not passing by value saves us from
the deep copy. This is the case in the prototype integration of user space
instrumentation as described in
http://go/stadia-orbit-user-space-instrumentation-and-unwinding#heading=h.a3i5cfh8pexm
In that prototype with the example used in the document (Trata's
`item_coordinate_frame::update`), CPU utilization of `PSSI::RcvEvents` is ~13%
instead of ~22%.
As this is a generic improvement, let's make it regardless of the prototype.

**R**-value reference because we signal the general intention of moving from the
parameter and leaving it in an undefined state, and because we want to accept
temporaries (again in line with moving from the parameter).

Bug: http://b/179969217, http://b/194704608

Test: Unit tests. Capture Trata with as many data collection features enabled as
possible.