Remove ModuleData::AddFunctionInfoWithBuildId
This method is a leftover from an old fix to capture loading from the old
format, code that is now gone.
It "pollutes" the interface of `ModuleData` as it adds a different way of
adding functions in addition to `AddSymbols`. Indeed, by bypassing `AddSymbols`,
it has the problem that it doesn't update `name_to_function_info_map_` and
`hash_to_function_map_`.
It's only used in a couple of tests as a shortcut, so let's just remove it.

Test: Run affected unit tests.