Add [[maybe_unused]] to write-only volatile vars (#2863)
In the UserSpaceInstrumentation tests we use volatile variables to force
some code generation. But recent clang trunk started to complain about
these variables being written only despite them being volatile. Since
volatile is such a complicated topic it's hard to judge if clang is
acting here correctly or not. Anyway, adding a `[[maybe_unused]]`
silences the warning and the code generation remains correct, so I
believe that is a viable quick fix.

If it turns out that the compiler is allowed to optimize away local
volatile variables we might need to find a different way in the future.
Global variables would be one option for example.