Optimize ScopeTree::Height() (#2866)
This PR is optimizing Height in ScopeTree. Since we are using this
structure in ThreadTrack, we should use Height() here to get the maximum
depth instead of timer_data.

So, in this PR we are optimizing Height() from O(n) to O(1), given that
the last depth should always have some nodes in the current
implementation.