Create ThreadTrackDataManager
This PR creates ThreadTrackDataManager, a temporal place (while
ThreadTrack is the only Track who uses ScopeTree) where the data inside
ThreadTrack will be stored. The class is really similar to
TimerDataManager but uses instead ScopeTreeTimerData.

The interface includes:

- CreateScopeTreeTimerData: To create a new ScopeTreeTimerData for a new
track. In the future might be a private function and all the logic
related will be modified by including AddNewTimer
(go/stadia-orbit-track-creation), who will use this
private function when needed.

- GetAllThreadTimerChains: Currently used in TimeGraph for looking for
previous/next function call and in some other places. Should be replaced
in the future by some queries.

- OnCaptureComplete: To build the ScopeTree when loading a capture.

This is a puzzle piece inside https://github.com/google/orbit/pull/2860,
aiming to solve http://b/202110929.

PS: I'm open to advices about the name of this class,
ScopeTreeTimerDataManager seems to be incorrect since it is being used
for ThreadTracks and we have a map based on thread_id. Actually, this
will help us to solve the multi-threading issue written in
go/stadia-orbit-track-creation for ThreadTracks.