Unify Depth, Height and MaxDepth in ScopeTree
In this PR we are unifying all this 3 methods with Depth(). So, depth in
a ScopeTree is 0-indexed, and the function Depth() will return 0 only if
there are no timers. MaxDepth was misleading.

This is a puzzle piece inside https://github.com/google/orbit/pull/2860,
aiming to solve http://b/202110929.