Compare inodes instead of file paths in tests (#2868)
When mapped files are symlinks (as it is the case in the internal
testing environment) or hardlinks the path that appears in the maps
file might be different and all the tests comparing file paths will
fail.

To avoid that problem this commit moves to comparing inodes and device ids
instead of file paths which is much more robust.

I also tried to fix the problem with `realpath` but that didn't help in
all cases. I'm not a 100% sure why but have been moving to inode
comparison anyway.