Refactor some functions in ScopeTreeTimerData
This small PR includes a few comments, a renaming (OnCaptureComplete),
and modified the return type from reference to pointer in
GetLeft/Right/Up/Down since the answer could be nullptr.

After this PR, there will be a ThreadTrackDataProvider who use this
class to answer some queries.

This is a puzzle piece inside #2860,
aiming to solve http://b/202110929.