Add ThreadTrackDataProvider
In this PR we are adding ThreadTrackDataProvider, which will provide
several queries for ThreadTrack to get information about their Timers
and the timers themselves.

A follow-up is coming where we will replace TimerData inside
ThreadTracks by a pointer to this provider and we will replace the
access to that data by the queries. All replacement are written of
https://github.com/google/orbit/pull/2860 (draft PR).

ThreadTrackDataProvider is using ThreadTrackDataManager to get the right
ScopeTreeTimerData and provides several methods:

- AddTimer: To add a timer and create the correspondent data if needed.
In the next PR will be called by the UI but in a follow-up will be
directly called by the CaptureThread outside the UI.

- GetAllThreadTimerChains: Used in the current code for jumping, will be
modified in the future.

- GetAllThreadId: Will be used in the future to create Tracks before
drawing instead of the current flow
(http://go/stadia-orbit-track-creation).

- GetTimers(thread_id, start, end): The main core of this class,
provides timers in a certain range.

- Metadata Queries(thread_id): IsEmpty, GetNumberOfTimers, GetMinTime, GetMaxTime,
GetDepth, used to get some data about the timers.

- GetLeft/Right/Up/Down: To be able to move with the arrow-keys from a
timer to another one.

- OnCaptureComplete(): Used internally to communicate that a Capture is
Complete to the data. Therefore, it can use a different structure given
that not new timers will appear.

This is a puzzle piece in solving http://b/202110929.