Remove world top left from viewport (#2842)
Bug: b/191231286

This finally moves all scrolling-functionality into Batcher and TextRenderer, and ownership of it into TimeGraph. I've decided to not generalize the functionality by introducing a scrollable container inheriting from CaptureViewElement as I don't expect the functionality to be used by other elements, and it would require to pass down all the mouse handling to CaptureViewElement as well and would overall just complicate things. So at least for now, ownership of the scrollbars still stays with CaptureWindow.

Test: Orbit builds and looks correct. Tested scrolling, vertical zoom, track type visibility, expanding and collapsing tracks. Also tested the corner-case where we scroll the window all the way down, then collapse a track.