Search in all modules for function name (#2874)
At some point the UserSpaceInstrumentation tests (TrampolineTest) need
to find a function's address range by name. This used to be done by
looking for that function in the main module (the executable's module).

This doesn't work in internal builds because they might be dynamically
linked (depending on the build config) which means the main module only
contains the GTest main function and some other GTest related code and
the actual unit tests are in a different shared library.

To make this work I changed the code such that the main module is
checked first (for performance reasons), but if the function is not
found the code falls back to searching through all the other modules.