Make CaptureFinished also a ProducerCaptureEvent for consistency (#2884)
Looking at `CaptureServiceImpl`, all events produced there (with
`kRootProducerId`) are `ProducerCaptureEvent`s, that then
`ProducerEventProcessor` makes `ClientCaptureEvent`s. The simplest example is
`CaptureStarted`.
But this wasn't the case for `CaptureFinished`, which was added directly to
`capture_event_buffer`, which is instead only supposed to be used by
`ProducerEventProcessor`. So we fix the inconsistency.

Test: Capture Trata.