Move SamplingReportDataView and add tests (#2872)
* Use AppInterface in SamplingReportDataView

With this change, we made SamplingReportDataView independent of
OrbitApp by chaning to use AppIterface.

* Add SamplingReportInterface

With this change, we added SamplingReportInterface. We made
SamplingReport to inherit from SamplingReportInterface and changed
SamplingReportDataView to use the interface class.

The purpose is to make SamplingReportDataView independent with the
OrbitGl module such that it can be moved to the DataViews module.

* Use SamplingReportInterface in SamplingReportDataView

With this change, we made SamplingReportDataView to use
SamplingReportInterface instead of SamplingReport.

* Move SamplingReportDataView to DataViews module

With this change, we moved the SamplingReportDataView to the DataViews
module.

* Add unit test for SamplingReportDataView

With this change, we add unit tests for the SamplingReportDataView.