CaptureEventBuffer+Sender performance rework: double buffering, Arenas (#2881)
Previously, `ProducerEventProcessor` was adding the finished
`ClientCaptureEvent`s into `SenderThreadCaptureEventBuffer`, implementation of
`CaptureEventBuffer`.
This was then periodically passing the `ClientCaptureEvent`s to
`GrpcCaptureEventSender`, instantiation of `CaptureEventSender`, which was then
assembling multiple `ClientCaptureEvent`s into `CaptureResponse`s and sending
them to the client.
`CaptureEventBuffer` and `CaptureEventSender` were separate for (maybe) clarity,
but the `CaptureEventSender` interface was not needed (only
`CaptureEventBuffer`).
The separation limited the possibility of optimization.

This reworks this structure for performance.
`SenderThreadCaptureEventBuffer` and `GrpcCaptureEventSender` are merged into a
single `GrpcClientCaptureEventCollector`, implementation of the interface
`ClientCaptureEventCollector` (basically renamed from `CaptureEventBuffer` as
it's not just a buffer anymore) that `ProducerEventProcessor` is aware of.
We optimize by adding the use of protobuf Arenas, which compared to before
prevent the allocation of `ClientCaptureEvent`s when assembling a
`CaptureResponse`. For `ClientCaptureEvent`s that were on the heap, this results
in copies that were previously moves, but saving memory allocations more than
compensates (see below).
We then introduce double buffering by using two Arenas. This is now possible
because we now have a single class, and we can already build `CaptureResponse`s
as `ClientCaptureEvent`s come in.

The original `SenderThreadCaptureEventBuffer` and `GrpcCaptureEventSender` were
in `CaptureServiceImpl.cpp`, but we move the new
`GrpcClientCaptureEventCollector` to its own file.

We also add some unit tests for the new class.

Performance was tested in two very different configurations:
1) Capturing Trata with scheduling, with sampling at 5000 Hz, and with
`world_sphere` dynamically instrumented;
2) Capturing AdventureUnity (in game) collecting only manual instrumentation.

In case 1, the CPU utilization of `SenderThread` goes down from 11% to 9%, with
the CPU utilization of `OrbitService` as a whole decreasing of more or less the
same amount (kind of hard to measure, ~86% to ~84%).
The improvement is limited because memory allocations are saved by the Arena,
but some moves from heap to heap became copies into an Arena. Either way,
because of what follows, it's already enough that performance stays at least the
same. (Note that in the future we should try to prevent more of the allocations
on the heap of the events we were previously moving from, hence saving more on
allocations while keeping the same copies.)
But where this change shines is in case 2, with the CPU utilization of
`SenderThread` going down from 19% to 11%, and CPU utilization of `OrbitService`
as a whole decreasing of more or less the same amount (44% to 35%).
The improvement is so significant here because events from manual
instrumentation already come from an Arena and were already copied, so in this
case we prevent memory allocations with no downsides.

Test: As part of the performance testing as described. Plus additional generic
testing capturing Trata and AdventureUnity.