Reduce memory allocations for PerfEvent’s. (#2892)
Previously PerfEvent’s were dynamically allocated when we read from the ring
buffers and passed around in unique_ptr’s. The specialisations of PerfEvent
are handed over to PerfEventProcessor which sorts them and finally consumes
them using a visitor pattern.

The problem with this is that a lot of heap allocations for the PerfEvents need
to happen.

This PR replaces the specialisations of PerfEvent by a std::variant and uses
std::visit to do the visiting of the events in PerfEventProcessor. This way the
events are stack allocated or emplaced into containers. So no “per event”
dynamic allocations on the heap are needed anymore (not entirely true: some
events contain unique_ptr’s with large chunks of data, e.g. stack samples).

Now the PerfEvents are passed around by value which is potentially more
expensive than passing around a pointer. But the positive effect of the avoided
allocation dominates. Depending on the use case there is a decrease of 15% to
32% in cpu utilisation of OrbitService.

Bug: http://b/200128809
Tests: Unit test, LinuxTracingIntegrationTests on gamelet and workstation, run
with trata/OrbitUI