Fix all known race conditions in the tests.
Sometimes the DexFiles tests fail in the memory size, so increase the
size slightly to avoid this problem.

Increase the timeout in the ThreadEntry from 5 to 10 seconds.
If running on a slower config, give a little more time.

Update the attach code to be a little more error tolerant.
Specifically, increase the time allowed for a quiesce. Also try
and kick the process if it gets into a group-stop state.

Display the Unwinder object error code when a failure occurs in
an UnwindTest.

Retry the thread unwinds if they timeout since this can happen on
slower systems.

Bug: 203224453

Test: Ran unit tests for hours on cuttlefish and flame hwasan config.
Change-Id: Iefa660007e067b521d45f59067f5603d63a5a1fa
(cherry picked from commit eab6451b1a114d15c07be52d3436df4eea2421c5)