Simplify visitor pattern and related cleanups.
Just call visitor->Visit instead of using the events 'Accept' method. So the
'Accept' methods can be deleted and this leads to a bunch of mostly mechanical
changes in the tests.

In order to keep the changes to the test limited the visitor is changed to
accept a const reference instead of a pointer.

This leads to another problem: the const_casts in PerfEventQueue::TopEvent
and PopEvent. Turns out the events are almost exclusively const. Only
StackSamplePerfEvent and CallchainSamplePerfEvent need to be modified in a
specific way. So the cleaner solution here is to make the two fields mutable
and use const references to move the events around.

Test: Unit test. Also manually deployed an profiled with the UI.